HomeScreen and Caching data completed

+++ app/src/main/java/com/suhail/tentwenty/data/Result.kt
@@ -31,7 +31,7 @@ data class Result(
     @SerializedName("video")
     val video: Boolean,
     @SerializedName("vote_average")
-    val voteAverage: Int,
+    val voteAverage: Double,
     @SerializedName("vote_count")
     val voteCount: Int
 )
